requirements.txt
streamlit
google-cloud-bigquery
db-dtypes
pandas
pytest